djangorest/requirements.txt
Django==3.1.2
djangorestframework==3.12.1

black>=24.3.0